+++ src/candid_rpc.rs
@@ -2,18 +2,14 @@ mod cketh_conversion;
 
 use async_trait::async_trait;
 use candid::Nat;
-use cketh_common::{
-    eth_rpc::{ProviderError, ValidationError},
-    eth_rpc_client::{
-        providers::{RpcApi, RpcService},
-        EthRpcClient as CkEthRpcClient, MultiCallError, RpcConfig, RpcTransport,
-    },
-    lifecycle::EthereumNetwork,
-};
+use cketh_common::eth_rpc_client::{EthRpcClient as CkEthRpcClient, MultiCallError, RpcTransport};
 use ethers_core::{types::Transaction, utils::rlp};
-use evm_rpc_types::{Hex, Hex32};
+use evm_rpc_types::{
+    Hex, Hex32, MultiRpcResult, ProviderError, RpcResult, RpcServices, ValidationError,
+};
 use ic_cdk::api::management_canister::http_request::{CanisterHttpRequestArgument, HttpResponse};
 
+use crate::candid_rpc::cketh_conversion::into_rpc_error;
 use crate::{
     accounting::get_http_request_cost,
     add_metric_entry,
@@ -23,10 +19,7 @@ use crate::{
     },
     http::http_request,
     providers::resolve_rpc_service,
-    types::{
-        MetricRpcHost, MetricRpcMethod, MultiRpcResult, ResolvedRpcService, RpcMethod, RpcResult,
-        RpcServices,
-    },
+    types::{MetricRpcHost, MetricRpcMethod, ResolvedRpcService, RpcMethod},
 };
 
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -35,17 +28,29 @@ struct CanisterTransport;
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
 impl RpcTransport for CanisterTransport {
-    fn resolve_api(service: &RpcService) -> Result<RpcApi, ProviderError> {
-        Ok(resolve_rpc_service(service.clone())?.api())
+    fn resolve_api(
+        service: &cketh_common::eth_rpc_client::providers::RpcService,
+    ) -> Result<cketh_common::eth_rpc_client::providers::RpcApi, cketh_common::eth_rpc::ProviderError>
+    {
+        use crate::candid_rpc::cketh_conversion::{
+            from_rpc_service, into_provider_error, into_rpc_api,
+        };
+        Ok(into_rpc_api(
+            resolve_rpc_service(from_rpc_service(service.clone()))
+                .map_err(into_provider_error)?
+                .api(),
+        ))
     }
 
     async fn http_request(
-        service: &RpcService,
+        service: &cketh_common::eth_rpc_client::providers::RpcService,
         method: &str,
         request: CanisterHttpRequestArgument,
         effective_response_size_estimate: u64,
-    ) -> RpcResult<HttpResponse> {
-        let service = resolve_rpc_service(service.clone())?;
+    ) -> Result<HttpResponse, cketh_common::eth_rpc::RpcError> {
+        use crate::candid_rpc::cketh_conversion::{from_rpc_service, into_provider_error};
+        let service =
+            resolve_rpc_service(from_rpc_service(service.clone())).map_err(into_provider_error)?;
         let cycles_cost = get_http_request_cost(
             request
                 .body
@@ -55,7 +60,9 @@ impl RpcTransport for CanisterTransport {
             effective_response_size_estimate,
         );
         let rpc_method = MetricRpcMethod(method.to_string());
-        http_request(rpc_method, service, request, cycles_cost).await
+        http_request(rpc_method, service, request, cycles_cost)
+            .await
+            .map_err(into_rpc_error)
     }
 }
 
@@ -68,81 +75,35 @@ fn check_services<T>(services: Vec<T>) -> RpcResult<Vec<T>> {
 
 fn get_rpc_client(
     source: RpcServices,
-    config: RpcConfig,
+    config: evm_rpc_types::RpcConfig,
 ) -> RpcResult<CkEthRpcClient<CanisterTransport>> {
-    Ok(match source {
-        RpcServices::Custom { chain_id, services } => CkEthRpcClient::new(
-            EthereumNetwork(chain_id),
-            Some(
-                check_services(services)?
-                    .into_iter()
-                    .map(RpcService::Custom)
-                    .collect(),
-            ),
-            config,
-        ),
-        RpcServices::EthMainnet(services) => CkEthRpcClient::new(
-            EthereumNetwork::MAINNET,
-            Some(
-                check_services(services.unwrap_or_else(|| DEFAULT_ETH_MAINNET_SERVICES.to_vec()))?
-                    .into_iter()
-                    .map(RpcService::EthMainnet)
-                    .collect(),
-            ),
-            config,
-        ),
-        RpcServices::EthSepolia(services) => CkEthRpcClient::new(
-            EthereumNetwork::SEPOLIA,
-            Some(
-                check_services(services.unwrap_or_else(|| DEFAULT_ETH_SEPOLIA_SERVICES.to_vec()))?
-                    .into_iter()
-                    .map(RpcService::EthSepolia)
-                    .collect(),
-            ),
-            config,
-        ),
-        RpcServices::ArbitrumOne(services) => CkEthRpcClient::new(
-            EthereumNetwork::ARBITRUM,
-            Some(
-                check_services(services.unwrap_or_else(|| DEFAULT_L2_MAINNET_SERVICES.to_vec()))?
-                    .into_iter()
-                    .map(RpcService::ArbitrumOne)
-                    .collect(),
-            ),
-            config,
-        ),
-        RpcServices::BaseMainnet(services) => CkEthRpcClient::new(
-            EthereumNetwork::BASE,
-            Some(
-                check_services(services.unwrap_or_else(|| DEFAULT_L2_MAINNET_SERVICES.to_vec()))?
-                    .into_iter()
-                    .map(RpcService::BaseMainnet)
-                    .collect(),
-            ),
-            config,
-        ),
-        RpcServices::OptimismMainnet(services) => CkEthRpcClient::new(
-            EthereumNetwork::OPTIMISM,
-            Some(
-                check_services(services.unwrap_or_else(|| DEFAULT_L2_MAINNET_SERVICES.to_vec()))?
-                    .into_iter()
-                    .map(RpcService::OptimismMainnet)
-                    .collect(),
-            ),
-            config,
-        ),
-    })
+    use crate::candid_rpc::cketh_conversion::{
+        into_ethereum_network, into_rpc_config, into_rpc_services,
+    };
+
+    let config = into_rpc_config(config);
+    let chain = into_ethereum_network(&source);
+    let providers = check_services(into_rpc_services(
+        source,
+        DEFAULT_ETH_MAINNET_SERVICES,
+        DEFAULT_ETH_SEPOLIA_SERVICES,
+        DEFAULT_L2_MAINNET_SERVICES,
+    ))?;
+    Ok(CkEthRpcClient::new(chain, Some(providers), config))
 }
 
 fn process_result<T>(method: RpcMethod, result: Result<T, MultiCallError<T>>) -> MultiRpcResult<T> {
+    use crate::candid_rpc::cketh_conversion::{from_rpc_error, from_rpc_service};
     match result {
         Ok(value) => MultiRpcResult::Consistent(Ok(value)),
         Err(err) => match err {
-            MultiCallError::ConsistentError(err) => MultiRpcResult::Consistent(Err(err)),
+            MultiCallError::ConsistentError(err) => {
+                MultiRpcResult::Consistent(Err(from_rpc_error(err)))
+            }
             MultiCallError::InconsistentResults(multi_call_results) => {
                 multi_call_results.results.iter().for_each(|(service, _)| {
                     if let Ok(ResolvedRpcService::Provider(provider)) =
-                        resolve_rpc_service(service.clone())
+                        resolve_rpc_service(from_rpc_service(service.clone()))
                     {
                         add_metric_entry!(
                             inconsistent_responses,
@@ -158,7 +119,15 @@ fn process_result<T>(method: RpcMethod, result: Result<T, MultiCallError<T>>) ->
                         )
                     }
                 });
-                MultiRpcResult::Inconsistent(multi_call_results.results.into_iter().collect())
+                MultiRpcResult::Inconsistent(
+                    multi_call_results
+                        .results
+                        .into_iter()
+                        .map(|(service, result)| {
+                            (from_rpc_service(service), result.map_err(from_rpc_error))
+                        })
+                        .collect(),
+                )
             }
         },
     }
@@ -169,7 +138,10 @@ pub struct CandidRpcClient {
 }
 
 impl CandidRpcClient {
-    pub fn new(source: RpcServices, config: Option<RpcConfig>) -> RpcResult<Self> {
+    pub fn new(
+        source: evm_rpc_types::RpcServices,
+        config: Option<evm_rpc_types::RpcConfig>,
+    ) -> RpcResult<Self> {
         Ok(Self {
             client: get_rpc_client(source, config.unwrap_or_default())?,
         })
@@ -287,11 +259,13 @@ fn get_transaction_hash(raw_signed_transaction_hex: &Hex) -> Option<Hex32> {
 #[cfg(test)]
 mod test {
     use super::*;
-    use cketh_common::eth_rpc::RpcError;
+    use crate::candid_rpc::cketh_conversion::into_rpc_service;
+    use evm_rpc_types::RpcError;
 
     #[test]
     fn test_process_result_mapping() {
-        use cketh_common::eth_rpc_client::{providers::EthMainnetService, MultiCallResults};
+        use cketh_common::eth_rpc_client::MultiCallResults;
+        use evm_rpc_types::{EthMainnetService, RpcService};
 
         let method = RpcMethod::EthGetTransactionCount;
 
@@ -303,7 +277,9 @@ mod test {
             process_result(
                 method,
                 Err(MultiCallError::<()>::ConsistentError(
-                    RpcError::ProviderError(ProviderError::MissingRequiredProvider)
+                    cketh_common::eth_rpc::RpcError::ProviderError(
+                        cketh_common::eth_rpc::ProviderError::MissingRequiredProvider
+                    )
                 ))
             ),
             MultiRpcResult::Consistent(Err(RpcError::ProviderError(
@@ -325,9 +301,12 @@ mod test {
             process_result(
                 method,
                 Err(MultiCallError::InconsistentResults(MultiCallResults {
-                    results: vec![(RpcService::EthMainnet(EthMainnetService::Ankr), Ok(5))]
-                        .into_iter()
-                        .collect(),
+                    results: vec![(
+                        into_rpc_service(RpcService::EthMainnet(EthMainnetService::Ankr)),
+                        Ok(5)
+                    )]
+                    .into_iter()
+                    .collect(),
                 }))
             ),
             MultiRpcResult::Inconsistent(vec![(
@@ -340,10 +319,15 @@ mod test {
                 method,
                 Err(MultiCallError::InconsistentResults(MultiCallResults {
                     results: vec![
-                        (RpcService::EthMainnet(EthMainnetService::Ankr), Ok(5)),
                         (
-                            RpcService::EthMainnet(EthMainnetService::Cloudflare),
-                            Err(RpcError::ProviderError(ProviderError::NoPermission))
+                            into_rpc_service(RpcService::EthMainnet(EthMainnetService::Ankr)),
+                            Ok(5)
+                        ),
+                        (
+                            into_rpc_service(RpcService::EthMainnet(EthMainnetService::Cloudflare)),
+                            Err(cketh_common::eth_rpc::RpcError::ProviderError(
+                                cketh_common::eth_rpc::ProviderError::NoPermission
+                            ))
                         )
                     ]
                     .into_iter()