+++ src/main.rs
@@ -1,7 +1,4 @@
 use candid::candid_method;
-
-use cketh_common::eth_rpc_client::providers::RpcService;
-use cketh_common::eth_rpc_client::RpcConfig;
 use cketh_common::logs::INFO;
 use evm_rpc::accounting::{get_cost_with_collateral, get_http_request_cost};
 use evm_rpc::candid_rpc::CandidRpcClient;
@@ -13,7 +10,13 @@ use evm_rpc::memory::{
 };
 use evm_rpc::metrics::encode_metrics;
 use evm_rpc::providers::{find_provider, resolve_rpc_service, PROVIDERS, SERVICE_PROVIDER_MAP};
-use evm_rpc::types::{Provider, ProviderId, RpcAccess, RpcResult};
+use evm_rpc::types::{Provider, ProviderId, RpcAccess};
+use evm_rpc::{
+    http::{json_rpc_request, transform_http_request},
+    memory::UNSTABLE_METRICS,
+    types::{InitArgs, MetricRpcMethod, Metrics},
+};
+use evm_rpc_types::{Hex32, MultiRpcResult, RpcResult};
 use ic_canister_log::log;
 use ic_canisters_http_types::{
     HttpRequest as AssetHttpRequest, HttpResponse as AssetHttpResponse, HttpResponseBuilder,
@@ -22,13 +25,6 @@ use ic_cdk::api::is_controller;
 use ic_cdk::api::management_canister::http_request::{HttpResponse, TransformArgs};
 use ic_cdk::{query, update};
 use ic_nervous_system_common::serve_metrics;
-
-use evm_rpc::{
-    http::{json_rpc_request, transform_http_request},
-    memory::UNSTABLE_METRICS,
-    types::{InitArgs, MetricRpcMethod, Metrics, MultiRpcResult, RpcServices},
-};
-use evm_rpc_types::Hex32;
 
 pub fn require_api_key_principal_or_controller() -> Result<(), String> {
     let caller = ic_cdk::caller();
@@ -42,8 +38,8 @@ pub fn require_api_key_principal_or_controller() -> Result<(), String> {
 #[update(name = "eth_getLogs")]
 #[candid_method(rename = "eth_getLogs")]
 pub async fn eth_get_logs(
-    source: RpcServices,
-    config: Option<RpcConfig>,
+    source: evm_rpc_types::RpcServices,
+    config: Option<evm_rpc_types::RpcConfig>,
     args: evm_rpc_types::GetLogsArgs,
 ) -> MultiRpcResult<Vec<evm_rpc_types::LogEntry>> {
     match CandidRpcClient::new(source, config) {
@@ -55,8 +51,8 @@ pub async fn eth_get_logs(
 #[update(name = "eth_getBlockByNumber")]
 #[candid_method(rename = "eth_getBlockByNumber")]
 pub async fn eth_get_block_by_number(
-    source: RpcServices,
-    config: Option<RpcConfig>,
+    source: evm_rpc_types::RpcServices,
+    config: Option<evm_rpc_types::RpcConfig>,
     block: evm_rpc_types::BlockTag,
 ) -> MultiRpcResult<evm_rpc_types::Block> {
     match CandidRpcClient::new(source, config) {
@@ -68,8 +64,8 @@ pub async fn eth_get_block_by_number(
 #[update(name = "eth_getTransactionReceipt")]
 #[candid_method(rename = "eth_getTransactionReceipt")]
 pub async fn eth_get_transaction_receipt(
-    source: RpcServices,
-    config: Option<RpcConfig>,
+    source: evm_rpc_types::RpcServices,
+    config: Option<evm_rpc_types::RpcConfig>,
     tx_hash: Hex32,
 ) -> MultiRpcResult<Option<evm_rpc_types::TransactionReceipt>> {
     match CandidRpcClient::new(source, config) {
@@ -81,8 +77,8 @@ pub async fn eth_get_transaction_receipt(
 #[update(name = "eth_getTransactionCount")]
 #[candid_method(rename = "eth_getTransactionCount")]
 pub async fn eth_get_transaction_count(
-    source: RpcServices,
-    config: Option<RpcConfig>,
+    source: evm_rpc_types::RpcServices,
+    config: Option<evm_rpc_types::RpcConfig>,
     args: evm_rpc_types::GetTransactionCountArgs,
 ) -> MultiRpcResult<evm_rpc_types::Nat256> {
     match CandidRpcClient::new(source, config) {
@@ -94,8 +90,8 @@ pub async fn eth_get_transaction_count(
 #[update(name = "eth_feeHistory")]
 #[candid_method(rename = "eth_feeHistory")]
 pub async fn eth_fee_history(
-    source: RpcServices,
-    config: Option<RpcConfig>,
+    source: evm_rpc_types::RpcServices,
+    config: Option<evm_rpc_types::RpcConfig>,
     args: evm_rpc_types::FeeHistoryArgs,
 ) -> MultiRpcResult<evm_rpc_types::FeeHistory> {
     match CandidRpcClient::new(source, config) {
@@ -107,8 +103,8 @@ pub async fn eth_fee_history(
 #[update(name = "eth_sendRawTransaction")]
 #[candid_method(rename = "eth_sendRawTransaction")]
 pub async fn eth_send_raw_transaction(
-    source: RpcServices,
-    config: Option<RpcConfig>,
+    source: evm_rpc_types::RpcServices,
+    config: Option<evm_rpc_types::RpcConfig>,
     raw_signed_transaction_hex: evm_rpc_types::Hex,
 ) -> MultiRpcResult<evm_rpc_types::SendRawTransactionStatus> {
     match CandidRpcClient::new(source, config) {
@@ -124,7 +120,7 @@ pub async fn eth_send_raw_transaction(
 #[update]
 #[candid_method]
 async fn request(
-    service: RpcService,
+    service: evm_rpc_types::RpcService,
     json_rpc_payload: String,
     max_response_bytes: u64,
 ) -> RpcResult<String> {
@@ -141,7 +137,7 @@ async fn request(
 #[query(name = "requestCost")]
 #[candid_method(query, rename = "requestCost")]
 fn request_cost(
-    _service: RpcService,
+    _service: evm_rpc_types::RpcService,
     json_rpc_payload: String,
     max_response_bytes: u64,
 ) -> RpcResult<u128> {
@@ -163,7 +159,7 @@ fn get_providers() -> Vec<Provider> {
 
 #[query(name = "getServiceProviderMap")]
 #[candid_method(query, rename = "getServiceProviderMap")]
-fn get_service_provider_map() -> Vec<(RpcService, ProviderId)> {
+fn get_service_provider_map() -> Vec<(evm_rpc_types::RpcService, ProviderId)> {
     SERVICE_PROVIDER_MAP.with(|map| map.iter().map(|(k, v)| (k.clone(), *v)).collect())
 }