+++ src/types.rs
@@ -1,9 +1,4 @@
 use candid::{CandidType, Principal};
-use cketh_common::eth_rpc::RpcError;
-use cketh_common::eth_rpc_client::providers::{
-    EthMainnetService, EthSepoliaService, L2MainnetService, RpcApi, RpcService,
-};
-
 use ic_cdk::api::management_canister::http_request::HttpHeader;
 use ic_stable_structures::{BoundedStorable, Storable};
 use serde::{Deserialize, Serialize};
@@ -26,12 +21,12 @@ pub struct InitArgs {
 }
 
 pub enum ResolvedRpcService {
-    Api(RpcApi),
+    Api(evm_rpc_types::RpcApi),
     Provider(Provider),
 }
 
 impl ResolvedRpcService {
-    pub fn api(&self) -> RpcApi {
+    pub fn api(&self) -> evm_rpc_types::RpcApi {
         match self {
             Self::Api(api) => api.clone(),
             Self::Provider(provider) => provider.api(),
@@ -268,27 +263,27 @@ pub struct Provider {
     #[serde(rename = "chainId")]
     pub chain_id: u64,
     pub access: RpcAccess,
-    pub alias: Option<RpcService>,
+    pub alias: Option<evm_rpc_types::RpcService>,
 }
 
 impl Provider {
-    pub fn api(&self) -> RpcApi {
+    pub fn api(&self) -> evm_rpc_types::RpcApi {
         match &self.access {
             RpcAccess::Authenticated { auth, public_url } => match get_api_key(self.provider_id) {
                 Some(api_key) => match auth {
-                    RpcAuth::BearerToken { url } => RpcApi {
+                    RpcAuth::BearerToken { url } => evm_rpc_types::RpcApi {
                         url: url.to_string(),
-                        headers: Some(vec![HttpHeader {
+                        headers: Some(vec![evm_rpc_types::HttpHeader {
                             name: "Authorization".to_string(),
                             value: format!("Bearer {}", api_key.read()),
                         }]),
                     },
-                    RpcAuth::UrlParameter { url_pattern } => RpcApi {
+                    RpcAuth::UrlParameter { url_pattern } => evm_rpc_types::RpcApi {
                         url: url_pattern.replace(API_KEY_REPLACE_STRING, api_key.read()),
                         headers: None,
                     },
                 },
-                None => RpcApi {
+                None => evm_rpc_types::RpcApi {
                     url: public_url
                         .unwrap_or_else(|| {
                             panic!(
@@ -300,7 +295,7 @@ impl Provider {
                     headers: None,
                 },
             },
-            RpcAccess::Unauthenticated { public_url } => RpcApi {
+            RpcAccess::Unauthenticated { public_url } => evm_rpc_types::RpcApi {
                 url: public_url.to_string(),
                 headers: None,
             },
@@ -352,148 +347,12 @@ pub enum RpcAuth {
     },
 }
 
-pub type RpcResult<T> = Result<T, RpcError>;
-
-#[derive(Clone, Debug, Eq, PartialEq, CandidType, Deserialize)]
-pub enum MultiRpcResult<T> {
-    Consistent(RpcResult<T>),
-    Inconsistent(Vec<(RpcService, RpcResult<T>)>),
-}
-
-impl<T> MultiRpcResult<T> {
-    pub fn map<R>(self, mut f: impl FnMut(T) -> R) -> MultiRpcResult<R> {
-        match self {
-            MultiRpcResult::Consistent(result) => MultiRpcResult::Consistent(result.map(f)),
-            MultiRpcResult::Inconsistent(results) => MultiRpcResult::Inconsistent(
-                results
-                    .into_iter()
-                    .map(|(service, result)| {
-                        (
-                            service,
-                            match result {
-                                Ok(ok) => Ok(f(ok)),
-                                Err(err) => Err(err),
-                            },
-                        )
-                    })
-                    .collect(),
-            ),
-        }
-    }
-
-    pub fn consistent(self) -> Option<RpcResult<T>> {
-        match self {
-            MultiRpcResult::Consistent(result) => Some(result),
-            MultiRpcResult::Inconsistent(_) => None,
-        }
-    }
-
-    pub fn inconsistent(self) -> Option<Vec<(RpcService, RpcResult<T>)>> {
-        match self {
-            MultiRpcResult::Consistent(_) => None,
-            MultiRpcResult::Inconsistent(results) => Some(results),
-        }
-    }
-
-    pub fn expect_consistent(self) -> RpcResult<T> {
-        self.consistent().expect("expected consistent results")
-    }
-
-    pub fn expect_inconsistent(self) -> Vec<(RpcService, RpcResult<T>)> {
-        self.inconsistent().expect("expected inconsistent results")
-    }
-}
-
-impl<T> From<RpcResult<T>> for MultiRpcResult<T> {
-    fn from(result: RpcResult<T>) -> Self {
-        MultiRpcResult::Consistent(result)
-    }
-}
-
-#[derive(Clone, CandidType, Deserialize)]
-pub enum RpcServices {
-    Custom {
-        #[serde(rename = "chainId")]
-        chain_id: u64,
-        services: Vec<RpcApi>,
-    },
-    EthMainnet(Option<Vec<EthMainnetService>>),
-    EthSepolia(Option<Vec<EthSepoliaService>>),
-    ArbitrumOne(Option<Vec<L2MainnetService>>),
-    BaseMainnet(Option<Vec<L2MainnetService>>),
-    OptimismMainnet(Option<Vec<L2MainnetService>>),
-}
-
 #[cfg(test)]
 mod test {
     use candid::Principal;
-    use cketh_common::{
-        eth_rpc::RpcError,
-        eth_rpc_client::providers::{EthMainnetService, RpcService},
-    };
     use ic_stable_structures::Storable;
 
-    use crate::types::{ApiKey, BoolStorable, MultiRpcResult, PrincipalStorable};
-
-    #[test]
-    fn test_multi_rpc_result_map() {
-        use cketh_common::eth_rpc::ProviderError;
-
-        let err = RpcError::ProviderError(ProviderError::ProviderNotFound);
-        assert_eq!(
-            MultiRpcResult::Consistent(Ok(5)).map(|n| n + 1),
-            MultiRpcResult::Consistent(Ok(6))
-        );
-        assert_eq!(
-            MultiRpcResult::Consistent(Err(err.clone())).map(|()| unreachable!()),
-            MultiRpcResult::Consistent(Err(err.clone()))
-        );
-        assert_eq!(
-            MultiRpcResult::Inconsistent(vec![(
-                RpcService::EthMainnet(EthMainnetService::Ankr),
-                Ok(5)
-            )])
-            .map(|n| n + 1),
-            MultiRpcResult::Inconsistent(vec![(
-                RpcService::EthMainnet(EthMainnetService::Ankr),
-                Ok(6)
-            )])
-        );
-        assert_eq!(
-            MultiRpcResult::Inconsistent(vec![
-                (RpcService::EthMainnet(EthMainnetService::Ankr), Ok(5)),
-                (
-                    RpcService::EthMainnet(EthMainnetService::Cloudflare),
-                    Ok(10)
-                )
-            ])
-            .map(|n| n + 1),
-            MultiRpcResult::Inconsistent(vec![
-                (RpcService::EthMainnet(EthMainnetService::Ankr), Ok(6)),
-                (
-                    RpcService::EthMainnet(EthMainnetService::Cloudflare),
-                    Ok(11)
-                )
-            ])
-        );
-        assert_eq!(
-            MultiRpcResult::Inconsistent(vec![
-                (RpcService::EthMainnet(EthMainnetService::Ankr), Ok(5)),
-                (
-                    RpcService::EthMainnet(EthMainnetService::PublicNode),
-                    Err(err.clone())
-                )
-            ])
-            .map(|n| n + 1),
-            MultiRpcResult::Inconsistent(vec![
-                (RpcService::EthMainnet(EthMainnetService::Ankr), Ok(6)),
-                (
-                    RpcService::EthMainnet(EthMainnetService::PublicNode),
-                    Err(err)
-                )
-            ])
-        );
-    }
+    use crate::types::{ApiKey, BoolStorable, PrincipalStorable};
 
     #[test]
     fn test_api_key_debug_output() {